Add files via upload
How it functions:
Through the functions we get through the "import datetime" we can use the commands it provides us to show our date and time in a simple manner. We fist name the date and time I used time_now. Then we make it a string using the commands we gain with the import. I named it timedate_format. The last thing to do is print it put what you want in "" and put the timedate_format at the end.

Instructions:
You would first download this file, go to a machine that has python install so you can use the commands it has available. And to get the desired result you would type python than the files name in the command prompt. So it would look something like this: python time_and_date_currently.py

This is the expected output:
('Date and Time Currently', 'Date: 10-09-2023 | Time: 06:21:54 PM')